third_party/blink/web_tests/platform/generic/fast/dom/Range/range-modifycontents-expected.txt

Test modification of contents of a range

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Start test 1
PASS r.startOffset is 0
PASS r.endOffset is 0
PASS r.startContainer is document
PASS r.endContainer is document
PASS r.collapsed is true
PASS r.commonAncestorContainer is document
Start test 2
PASS r.startOffset is 0
PASS r.endOffset is 1
PASS r.startContainer is document
PASS r.endContainer is document
PASS r.collapsed is false
PASS r.commonAncestorContainer is document
Start test 3
PASS r.startOffset is 0
PASS r.endOffset is 0
PASS r.startContainer is document
PASS r.endContainer is document
PASS r.collapsed is true
PASS r.commonAncestorContainer is document
Start test 4
PASS r.startOffset is 0
PASS r.endOffset is 3
PASS r.startContainer is c
PASS r.endContainer is c
PASS r.collapsed is false
PASS r.commonAncestorContainer is c
Start test 5
PASS r.startOffset is 0
PASS r.endOffset is 0
PASS r.startContainer is c
PASS r.endContainer is c
PASS r.collapsed is true
PASS r.commonAncestorContainer is c
Start test 6
PASS r.startOffset is 0
PASS r.endOffset is 0
PASS r.startContainer is c
PASS r.endContainer is c
PASS r.collapsed is true
PASS r.commonAncestorContainer is c
Start test 7
PASS r.startOffset is 0
PASS r.endOffset is 1
PASS r.startContainer is c
PASS r.endContainer is c
PASS r.collapsed is false
PASS r.commonAncestorContainer is c
Start test 8
PASS r.startOffset is 0
PASS r.endOffset is 0
PASS r.startContainer is c
PASS r.endContainer is c
PASS r.collapsed is true
PASS r.commonAncestorContainer is c
PASS successfullyParsed is true

TEST COMPLETE
